requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# Language Processing
langdetect>=1.0.9

# Data Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Additional utilities
typing-extensions>=4.0.0
